+++ src/SeExpr2/ExprBuiltins.cpp
@@ -266,7 +266,7 @@ Vec3d midhsi(int n, const Vec3d* args) {
 
         // scale hsi values according to mask (both directions)
         h *= m;
-        float absm = fabs(m);
+        float absm = fabs(static_cast<float>(m));
         s = s * absm + 1 - absm;
         i = i * absm + 1 - absm;
         if (m < 0) {
@@ -671,7 +671,7 @@ double fbm4(int n, const Vec3d* args) {
         case 3:
             octaves = int(clamp(args[2][0], 1, 8));
         case 2:
-            time = args[1][0];
+            time = static_cast<float>(args[1][0]);
         case 1:
             p = args[0];
     }
@@ -706,7 +706,7 @@ Vec3d vfbm4(int n, const Vec3d* args) {
         case 3:
             octaves = int(clamp(args[2][0], 1, 8));
         case 2:
-            time = args[1][0];
+            time = static_cast<float>(args[1][0]);
         case 1:
             p = args[0];
     }
@@ -885,7 +885,7 @@ Vec3d voronoiFn(VoronoiPointData& data, int n, const Vec3d* args) {
         case 4:
             return f2 - f1;
         case 5: {
-            float scalefactor = (pos2 - pos1).length() / ((pos1 - p).length() + (pos2 - p).length());
+            float scalefactor = static_cast<float>((pos2 - pos1).length() / ((pos1 - p).length() + (pos2 - p).length()));
             return smoothstep(f2 - f1, 0, 0.1 * scalefactor);
         }
     }
@@ -952,7 +952,7 @@ Vec3d cvoronoiFn(VoronoiPointData& data, int n, const Vec3d* args) {
         case 4:
             return (f2 - f1) * color;
         case 5: {
-            float scalefactor = (pos2 - pos1).length() / ((pos1 - p).length() + (pos2 - p).length());
+            float scalefactor = static_cast<float>((pos2 - pos1).length() / ((pos1 - p).length() + (pos2 - p).length()));
             return smoothstep(f2 - f1, 0, 0.1 * scalefactor) * color;
         }
     }
@@ -1110,7 +1110,7 @@ Vec3d rotate(int n, const Vec3d* args) {
     if (n != 3) return 0.0;
     const Vec3d& P = args[0];
     const Vec3d& axis = args[1];
-    float angle = args[2][0];
+    float angle = static_cast<float>(args[2][0]);
     double len = axis.length();
     if (!len) return P;
     return P.rotateBy(axis / len, angle);
@@ -1500,16 +1500,16 @@ class PrintFuncX : public ExprFuncSimple {
                 delete data;
                 assert(false);
             } else if (format[percentStart + 1] == '%') {
-                searchStart = percentStart + 2;
+                searchStart = static_cast<int>(percentStart + 2);
                 continue;
             } else if (format[percentStart + 1] == 'v' || format[percentStart + 1] == 'f') {
                 char c = format[percentStart + 1];
                 int code = (c == 'v') ? -1 : -2;
                 needed++;
-                if (bakeStart != percentStart) ranges.push_back(std::pair<int, int>(bakeStart, percentStart));
+                if (bakeStart != percentStart) ranges.push_back(std::pair<int, int>(bakeStart, static_cast<int>(percentStart)));
                 ranges.push_back(std::pair<int, int>(code, code));
                 items++;
-                searchStart = percentStart + 2;
+                searchStart = static_cast<int>(percentStart + 2);
                 bakeStart = searchStart;
             } else {
                 // node->addError("Invalid format string, only %v is allowed");
@@ -1520,7 +1520,7 @@ class PrintFuncX : public ExprFuncSimple {
                 assert(false);
             }
         }
-        if (bakeStart != format.length()) ranges.push_back(std::pair<int, int>(bakeStart, format.length()));
+        if (bakeStart != format.length()) ranges.push_back(std::pair<int, int>(bakeStart, static_cast<int>(format.length())));
 
         if (items != args.nargs() - 1) {
             // node->addError("Wrong number of arguments for format string");
